csm-logger: Add flag to check for first trigger
Time sync on actual device takes few seconds.
And once the time is changed the first run of this service
will not be able to fetch logs from the start i.e,logs with the
time before the sync wont be fetched by journald because we are
using --since option. Added a mechanism to differentiate first
trigger and subsequent triggers of the timer.

Change-Id: I1cd42e5a1b1fe2116c228be9a99fac5977de0c51